Implement client-server authentication flow
To get the app token to access the API, the client must first send the
bare minimum of a Facebook token and uid to /authenticate. A successful
authentication will return the app token which must be passed to the
HTTP header Authorization: Token token="...". A newly registered user
will return HTTP code 201, and a returning user will return 200.