Complete revision of database migrations and models
Relies more on associations to create and query records. This reduces
the complexity of the database schema, and will also allow for easier
querying by avoiding the need for complex joins.

API will of course need a complete revision also, but structure may
be adapted to extract business logic into separate controllers.
Looking into this.