Implemented message exchange mechanism (basic)
This is not complete, as the next scene is not being triggered on
app server. This will necessitate the use of GCM at some point. However,
basic logic until next scene triggering should be more or less up to
scratch. WRT routing, this is logically assigned to POST 'messages/:message_id'.

API keys are now generated and associated with a user upon creation of
the user record. Since this is essentially the only time when API keys
are being generated (at least for now), this is more to do with where
this responsibility is assigned (user model).

Minor terminology revision across app server to reinforce the distinction
between character and user messages (see database migration or schema
changes).

NOTE: At some point, the 'auth' resource has been changed to 'users' with
authentication done through POST 'users/:fb_user_id'. This is mainly to do
with logical structuring of routes, and the API's RESTfulness.