Revised APIs to use instance variables
This has worked very well as the instance variables can be set in the
after_validation block, it's associations used to perform error checks,
and then used in the business logic of the called route as well.

Helpers have also been reworked to remove now redundant helpers, and alter
existing ones to use ActiveRecord objects instead of their ids.