Revised models
Added delay to messages.
Note that the 'messages/{message_id}' PUT route is most definitely broken.
Will need to be fixed, but as it currently stands, it required too many changes
to the models, and is doing too much work anyway. By keeping the model associations
clean, all other routes should be working.